% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_stash.R
\name{clear_stash}
\alias{clear_stash}
\title{Clear the stash}
\usage{
clear_stash()
}
\value{
Returns \code{NULL} (invisibly).
}
\description{
Clears the hidden '.mustashe' directory.
}
\examples{
clear_stash()

}
