% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_mustashe.R
\name{config_mustashe}
\alias{config_mustashe}
\title{Configure default options for 'mustashe'}
\usage{
config_mustashe(use_here, verbose, functional)
}
\arguments{
\item{use_here}{Use the 'here' package when writing stashes to file. The
default behavior is to not use 'here.'}

\item{verbose}{The default value of \code{verbose} arguments when stashing or
un-stashing. The default behavior is \code{verbose = TRUE}.}

\item{functional}{The default value of \code{functional} arguments when
stashing. The default behavior is \code{functional = FALSE}.}
}
\description{
Sets options that affect the behavior of 'mustashe' or the default values for
stashing arguments. This function can be added to a .Rprofile to set the
defaults for 'mustashe' in a project.
}
\note{
Providing no value for any of the arguments results in no value being
  set. Therefore, \code{config_mustashe()} can be called with different
  arguments without over-writing previously set options.
}
