% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_here.R
\name{dont_use_here}
\alias{dont_use_here}
\title{Stop using the 'here' package when writing stashes to file (deprecated)}
\usage{
dont_use_here(silent = FALSE)
}
\arguments{
\item{silent}{A logical to silence the message from the function. (default
\code{FALSE})}
}
\value{
Returns \code{NULL} (invisibly).
}
\description{
Stop using the \code{here::here()} function from the
\href{https://here.r-lib.org}{'here'} package to create the file paths for
stashed objects.
}
\note{
This function has been deprecated and the use of the 'here' function
  should instead be set using \code{config_mustashe(use_here = FALSE)}.
}
\examples{
dont_use_here()
}
