% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stash.R
\name{stash}
\alias{stash}
\title{Stash an object.}
\usage{
stash(var, code, depends_on = NULL)
}
\arguments{
\item{var}{A variable to stash (as a string).}

\item{code}{The code to generate the object to be stashed.}

\item{depends_on}{A vector of other objects that this one depends on. Changes
to these objects will cause the re-running of the code, next time.}
}
\value{
\code{NULL}
}
\description{
Stash an object.
}
