% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unstash.R
\name{unstash}
\alias{unstash}
\title{Unstash an object}
\usage{
unstash(var)
}
\arguments{
\item{var}{The name or a vector of names of objects to remove.}
}
\value{
\code{NULL}
}
\description{
Remove an object from the stash.
}
\examples{
\dontrun{
stash("x",
{
    x <- 1
})

unstash("x")
}

}
