% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mustashe.R
\docType{package}
\name{mustashe}
\alias{mustashe}
\title{mustashe: Stash and Load Objects in R}
\description{
A simple system for saving and loading objects in R. Long running
computations can be stashed after the first run and then reloaded the next
time. Dependencies can be added to ensure that a computation is re-run if
any of its dependencies or inputs have changed.
}
