% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unstash.R
\name{unstash}
\alias{unstash}
\title{Unstash an object}
\usage{
unstash(var, single_var, verbose = TRUE)
}
\arguments{
\item{var}{The name or a vector of names of objects to remove.}

\item{single_var}{Specifies a single name (key) to remove. Use this
when the object was stashed using an arbitrary object as the key.}

\item{verbose}{Whether to print action statements (default TRUE).}
}
\value{
Returns \code{NULL} (invisibly).
}
\description{
Remove one or more objects from the stash.
}
\examples{
\donttest{
stash("x", {
  x <- 1
})

unstash("x")

stash(list(letters,cars), {7}, functional=TRUE)
unstash(single_var=list(letters,cars))

#' # Remove directory for this example - do not do in real use.
unlink(".mustashe", recursive = TRUE)
}

}
